fix(tests): add proper mocking for CI compatibility
Critical fix: VSCode tests will now pass in CI without requiring ARO binary or file system access.

Changes:
- Extract validateAroPath() to src/validation.ts (separate from vscode module)
- Add test dependencies: mocha, sinon, proxyquire, ts-node
- Create .mocharc.json for test configuration
- Add npm test script to package.json
- Implement proper mocking using proxyquire for Valid Paths tests:
  * Mock fs/promises.realpath to return path as-is
  * Mock fs/promises.access to succeed without file checks
  * Mock child_process.execFile to return "aro version 1.0.0"
  * Mock util.promisify to properly wrap callback-based functions

Test results: All 17 tests passing
- 9 security tests (no mocking needed - test validation logic directly)
- 3 valid path tests (mocked for CI - test with simulated file system)
- 5 version pattern tests (regex only - no mocking needed)

Fixes critical issue identified in PR review where tests would fail in CI
due to missing ARO binary and file paths.
