+++ Editor/vscode-aro/.mocharc.json
@@ -1,0 +1,6 @@
+{
+  "require": ["ts-node/register"],
+  "extensions": ["ts"],
+  "spec": "src/test/**/*.test.ts",
+  "recursive": true
+}
