+++ Editor/vscode-aro/package.json
@@ -26,8 +26,13 @@
     "languages": [
       {
         "id": "aro",
-        "aliases": ["ARO", "aro"],
-        "extensions": [".aro"],
+        "aliases": [
+          "ARO",
+          "aro"
+        ],
+        "extensions": [
+          ".aro"
+        ],
         "configuration": "./language-configuration.json"
       }
     ],
@@ -156,14 +161,22 @@
   "scripts": {
     "vscode:prepublish": "npm run compile",
     "compile": "tsc -p ./",
-    "watch": "tsc -watch -p ./"
+    "watch": "tsc -watch -p ./",
+    "test": "mocha -r ts-node/register src/test/**/*.test.ts"
   },
   "dependencies": {
     "vscode-languageclient": "^9.0.1"
   },
   "devDependencies": {
-    "@types/vscode": "^1.74.0",
+    "@types/mocha": "^10.0.0",
     "@types/node": "^18.0.0",
+    "@types/proxyquire": "^1.3.31",
+    "@types/sinon": "^17.0.0",
+    "@types/vscode": "^1.74.0",
+    "mocha": "^10.0.0",
+    "proxyquire": "^2.1.3",
+    "sinon": "^17.0.0",
+    "ts-node": "^10.0.0",
     "typescript": "^5.0.0"
   }
 }
